+++ src/containers/admins/Test/index.js
@@ -1,0 +1,19 @@
+import React from 'react';
+
+class AdminTestPage extends React.PureComponent {
+  state = {
+    test: '',
+  }
+
+  render() {
+    const { test } = this.state;
+    return (
+      <div>
+        this is test admin page
+        {test}
+      </div>
+    );
+  }
+}
+
+export default AdminTestPage;
