+++ src/modules/rootReducer/index.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import admin from 'containers/Login/reducer';
+import admin from 'containers/admins/Login/reducer';
 
 export default combineReducers({
   admin,
