+++ src/modules/rootSaga/index.js
@@ -1,5 +1,5 @@
 import { all } from 'redux-saga/effects';
-import { watchOnLogin } from 'containers/Login/saga';
+import { watchOnLogin } from 'containers/admins/Login/saga';
 
 export default function* rootSaga() {
   yield all([
